% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.R
\name{am}
\alias{am}
\title{Adaptive Metropolis algorithm}
\usage{
am(prior, pdf, t, d)
}
\arguments{
\item{prior}{A function(N,d) that draws N samples from a d-variate prior distribution.
Returns an N-by-d matrix.}

\item{pdf}{A function(prior) that calculates the density of the target distribution for given prior.
Returns an N-variate vector.}

\item{t}{\code{numeric}. Number of samples from the Markov chain.}

\item{d}{\code{numeric}. Number of parameters.}
}
\value{
\code{list} with named elements 'chain': a t-by-d matrix of parameter realisations for each iteration
of the Markov chain; and 'density': a t-variate vector of densities computed by \code{pdf} at each iteration
of the Markov chain.
}
\description{
Adaptive Metropolis algorithm
}
\references{
Code from 'Algorithm 2' of:

Vrugt, J. A.: "Markov chain Monte Carlo simulation using the DREAM software package:
Theory, concepts, and MATLAB implementation." Environmental Modelling & Software, 2016, 75, 273 -- 316,
\url{http://dx.doi.org/10.1016/j.envsoft.2015.08.013}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
