% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mc.R
\name{de_mc}
\alias{de_mc}
\title{Differential Evolution Markov Chain (DE-MC) algorithm}
\usage{
de_mc(prior, pdf, nc, t, d)
}
\arguments{
\item{prior}{A function(N,d) that draws N samples from a d-variate prior distribution.
Returns an N-by-d matrix.}

\item{pdf}{A function(prior) that calculates the density of the target distribution for given prior.
Returns an N-variate vector.}

\item{nc}{\code{numeric}. Number of chains evolved in parallel.}

\item{t}{\code{numeric}. Number of samples from the Markov chain.}

\item{d}{\code{numeric}. Number of parameters.}
}
\value{
\code{list} with named elements 'chain': a t-by-d-by-nc array of parameter realisations for each iteration
and Markov chain; and 'density': a t-by-nc matrix of densities computed by \code{pdf} at each iteration for each
Markov chain.
}
\description{
Differential Evolution Markov Chain (DE-MC) algorithm
}
\references{
Code from 'Algorithm 4' of:

Vrugt, J. A.: "Markov chain Monte Carlo simulation using the DREAM software package:
Theory, concepts, and MATLAB implementation." Environmental Modelling & Software, 2016, 75, 273 -- 316,
\url{http://dx.doi.org/10.1016/j.envsoft.2015.08.013}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
