% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{game}
\alias{game}
\title{Gaussian Mixture Importance (GAME) sampling}
\usage{
game(theta, theta_post, m_sub = NULL, m0 = NULL, m1, Jmax = 5, ic, omega,
  ob = FALSE, R = NULL, par.info = NULL, lik = NULL, fun = NULL, ...,
  obs = NULL, abc_rho = NULL, abc_e = NULL, glue_shape = NULL,
  lik_fun = NULL, ncores = 1)
}
\arguments{
\item{theta}{\code{numeric}. An m-by-d matrix of posterior samples (e.g. generated by
\code{\link[HydroBayes]{dream_parallel}}).}

\item{theta_post}{\code{numeric}. An m-dimensional vector of posterior densities for
\code{theta} (e.g. generated by \code{\link[HydroBayes]{dream_parallel}}).}

\item{m_sub}{\code{integer}. The number of sub-samples from m in \code{theta} and
\code{theta_post} to be drawn to speed-up estimation of the Gaussian mixture distribution.
This sub-sample of m will be excluded when drawing the \code{m1} samples. Must be
\code{<= nrow(theta)-m1}. Default: NULL (all m samples are used).}

\item{m0}{\code{integer}. Number of samples to be drawn from the mixture distribution
on wich the target function will be evaluated (an algorithmic parameter; only needed if \code{omega > 0};
also defines the number of times, \code{fun} needs to be evaluated and, thus, influences
the computing time, but also the sampling accuracy).}

\item{m1}{\code{integer}. Number of sub-samples to be drawn from \code{theta} to
sample the model evidence. As such, it needs to be \code{<= nrow(theta)}.}

\item{Jmax}{\code{integer}. Maximum number of components to be used for the Gaussian
mixture distribution which is used as catalyst to sample the model evidence. Volpi
et al. (2017) suggest a value of 5 (default).}

\item{ic}{\code{character} string defining the Information Criterion to be used to
evaluate the calibration of the Gaussian mixture distribution. One of: 'bic' -
Bayesian Information Criterion; 'var' - variance of the ratio between the density of
the target distribution and the density of its approximation (the Gaussian mixture
distribution).}

\item{omega}{\code{numeric}. Parameter defining the sampling procedure to be applied:
0: reciprocal importance sampling (RIS); 0 < omega < 1: bridge sampling with geometric
bridge (GB); 1: importance sampling (IS). Nedds to be within [0,1].}

\item{ob}{\code{logical}. If \code{TRUE}, bridge sampling with optimal bridge using
an initial value from GB or IS (depending on \code{omega}) will be applied. Default:
\code{FALSE}. If \code{omega == 0}, it will be ignored.}

\item{R}{\code{integer}. Number of iterations within bridge sampling with optimal
bridge. If \code{ob == TRUE} and \code{is.null(R)} (default), it will be set to
a value of 10 (suggested by Volpi et al., 2017). If \code{ob == FALSE}, it will be
ignored.}

\item{par.info}{\code{list}. See \code{\link[HydroBayes]{dream_parallel}}.}

\item{lik}{\code{integer}. See \code{\link[HydroBayes]{dream_parallel}}.}

\item{fun}{\code{character}. See \code{\link[HydroBayes]{dream_parallel}}.}

\item{...}{Additional arguments to \code{fun}, see \code{\link[HydroBayes]{dream_parallel}}.}

\item{obs}{\code{numeric}. See \code{\link[HydroBayes]{dream_parallel}}.}

\item{abc_rho}{\code{character}. See \code{\link[HydroBayes]{dream_parallel}}.}

\item{abc_e}{\code{numeric}. See \code{\link[HydroBayes]{dream_parallel}}.}

\item{glue_shape}{\code{numeric}. See \code{\link[HydroBayes]{dream_parallel}}.}

\item{lik_fun}{\code{character}. See \code{\link[HydroBayes]{dream_parallel}}.}

\item{ncores}{\code{integer}. See \code{\link[HydroBayes]{dream_parallel}}.}
}
\value{
A named \code{list} with the following elements:

\emph{Z}: Single value of the calculated marginal likelihood / Bayesian model evidence.

\emph{J}: Number of components of the best calibrated Gaussian mixture distribution
(\code{<= Jmax}).

\emph{ic_val}: Value of the selected Information Criterion of the best Gaussian mixture distribution.
}
\description{
Implementation of the GAME algorithm for sampling of the marginal likelihood,
also called model evidence, for use in Bayesian model comparison and/or averaging
frameworks.
}
\details{
Notation:

- m: Number of MC samples.\cr
- d: number of parameters.\cr
}
\note{
The arguments specific to function \code{\link[HydroBayes]{dream_parallel}}
are only needed if \code{omega > 0}, i.e. if additional evaluations of the target
function (\code{fun}) are necessary. Therefore, RIS might be most CPU efficient
(especially if an evaluation of \code{fun} requires a lot of computation time),
but could be inappropriate in sampling the model evidence!
}
\references{
Volpi, E., G. Schoups, G. Firmani and J. A. Vrugt: "Sworn testimony of the model
evidence: Gaussian Mixture Importance (GAME) sampling". Water Resources Research,
2017, 53, \url{http://dx.doi.org/10.1002/2016WR020167}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
