% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mixture.R
\name{mixture}
\alias{mixture}
\title{1-D mixture distribution as test pdf}
\usage{
mixture(x)
}
\arguments{
\item{x}{\code{numeric}. Value for which the function will be evaluated.}
}
\value{
Density of the mixture distribution for candidate value \code{x}.
}
\description{
1-D mixture distribution as test pdf
}
\references{
Based on Matlab code from Appendix C of:

Vrugt, J. A.: "Markov chain Monte Carlo simulation using the DREAM software package:
Theory, concepts, and MATLAB implementation." Environmental Modelling & Software, 2016, 75, 273 -- 316,
\url{http://dx.doi.org/10.1016/j.envsoft.2015.08.013}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
