% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelman_rubin.R
\name{R_stat}
\alias{R_stat}
\title{Gelman-Rubin convergence diagnostic}
\usage{
R_stat(x)
}
\arguments{
\item{x}{\code{numeric} array of [no. of samples]-by-[no. of parameters]-by-[no. of chains] dimensions}
}
\value{
A [no. of parameters]-dimensional vector of Gelman-Rubin diagnostic values.
}
\description{
Calculates diagnostic values to determine convergence of a MCMC algorithm.
}
\details{
The Gelman-Rubin value should be <= 1.2 for all parameters to declare convergence.
Otherwise the number of samples should be increased.
}
\references{
Algorithm based on Gelman and Rubin (1992) obtained from (Eqs. 32-35):

Vrugt, J. A.: "Markov chain Monte Carlo simulation using the DREAM software package:
Theory, concepts, and MATLAB implementation." Environmental Modelling & Software, 2016, 75, 273 -- 316,
\url{http://dx.doi.org/10.1016/j.envsoft.2015.08.013}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
