% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwm.R
\name{rwm}
\alias{rwm}
\title{Random Walk Metropolis algorithm}
\usage{
rwm(prior, pdf, t, d)
}
\arguments{
\item{prior}{A function that draws N samples from a d-variate prior distribution.}

\item{pdf}{A function that calculates the density of the target distribution for given parameters.}

\item{t}{\code{numeric}. Number of samples from the Markov chain.}

\item{d}{\code{numeric}. Number of parameters.}
}
\value{
\code{list} with named elements 'chain': a t x d matrix of parameter realisations for each iteration
of the Markov chain; and 'density': a t-variate vector of densities computed by \code{pdf} at each iteration
of the Markov chain.
}
\description{
Random Walk Metropolis algorithm
}
\references{
Code from 'Algorithm 1' of:

Vrugt, J. A.: "Markov chain Monte Carlo simulation using the DREAM software package:
Theory, concepts, and MATLAB implementation." Environmental Modelling & Software, 2016, 75, 273 -- 316,
\url{http://dx.doi.org/10.1016/j.envsoft.2015.08.013}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
