% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_t_distribution.R
\name{t_distribution}
\alias{t_distribution}
\title{Multivariate t-/Student distribution with 60 deg. of freedom}
\usage{
t_distribution(x)
}
\arguments{
\item{x}{\code{matrix} of type \code{numeric}. Rows correspond to observations and columns
to variables defining the dimensionality of the problem.}
}
\value{
A vector of type \code{numeric} and length equal to the number of rows of \code{x}.
}
\description{
Multivariate t-/Student distribution with 60 deg. of freedom
}
\references{
Based on Matlab code from Appendix C of:

Vrugt, J. A.: "Markov chain Monte Carlo simulation using the DREAM software package:
Theory, concepts, and MATLAB implementation." Environmental Modelling & Software, 2016, 75, 273 -- 316,
\url{http://dx.doi.org/10.1016/j.envsoft.2015.08.013}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
