% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream.R
\name{dream}
\alias{dream}
\title{Differential Evolution Adaptive Metropolis (DREAM) algorithm}
\usage{
dream(prior, pdf, nc, t, d, burnin = 0.1, delta = 3, c_val = 0.1,
  c_star = 1e-06, nCR = 3, p_g = 0.2)
}
\arguments{
\item{prior}{A function(N,d) that draws N samples from a d-variate prior distribution.
Returns an N-by-d matrix.}

\item{pdf}{A function(prior) that calculates the density of the target distribution for given prior.
Returns an N-variate vector.}

\item{nc}{\code{numeric}. Number of chains evolved in parallel.}

\item{t}{\code{numeric}. Number of samples from the Markov chain.}

\item{d}{\code{numeric}. Number of parameters.}

\item{burnin}{\code{numeric}. Length of the burn-in period as portion of t (\code{burnin period = burnin * t}).
Must be < 1! Default: 0.1.}

\item{delta}{\code{integer}. Maximum number of chain pairs used to generate the jump (default: 3).}

\item{c_val}{\code{numeric}. Lambda value is sampled from U[-c_val,c_val] (default: 0.3).}

\item{c_star}{\code{numeric}. Zeta value sampled from N[0,c_star]. Should be small compared to target
(i.e. in this case the normal) distribution. Default: 1e-6.}

\item{nCR}{\code{integer}. Length of vector with crossover probabilities for parameter subspace sampling (default: 3).}

\item{p_g}{\code{numeric}. Probability for gamma, the jump rate, being equal to 1. Default: 0.2.}
}
\value{
\code{list} with named elements 'chain': a t-by-d-by-nc array of parameter realisations for each iteration
and Markov chain; and 'density': a t-by-nc matrix of densities computed by \code{pdf} at each iteration for each
Markov chain.
}
\description{
Differential Evolution Adaptive Metropolis (DREAM) algorithm
}
\details{
To understand the notation (e.g. what is lambda, nCR etc.), have a look at Sect. 3.3
of the reference paper (see below).
}
\references{
Code from 'Algorithm 5' of:

Vrugt, J. A.: "Markov chain Monte Carlo simulation using the DREAM software package:
Theory, concepts, and MATLAB implementation." Environmental Modelling & Software, 2016, 75, 273 -- 316,
\url{http://dx.doi.org/10.1016/j.envsoft.2015.08.013}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
