% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream.R
\name{dream}
\alias{dream}
\title{Differential Evolution Adaptive Metropolis (DREAM) algorithm}
\usage{
dream(prior, pdf, nc, t, d, burnin = 0.1, delta = 3, c_val = 0.1,
  c_star = 1e-06, nCR = 3, p_g = 0.2, beta0 = 1, thin = 1,
  ncores = 1, verbose = TRUE)
}
\arguments{
\item{prior}{A function(N,d) that draws N samples from a d-variate prior distribution.
Returns an N-by-d matrix.}

\item{pdf}{A function(prior) that calculates the density of the target distribution for given prior.
Returns an N-variate vector.}

\item{nc}{\code{numeric}. Number of chains evolved in parallel.}

\item{t}{\code{numeric}. Number of samples from the Markov chain.}

\item{d}{\code{numeric}. Number of parameters.}

\item{burnin}{\code{numeric}. Length of the burn-in period as portion of t (\code{burnin period = burnin * t}).
Must be < 1! Default: 0.1.}

\item{delta}{\code{integer}. Maximum number of chain pairs used to generate the jump (default: 3).}

\item{c_val}{\code{numeric}. Lambda value is sampled from U[-c_val,c_val] (default: 0.3).}

\item{c_star}{\code{numeric}. Zeta value sampled from N[0,c_star]. Should be small compared to target
(i.e. in this case the normal) distribution. Default: 1e-6.}

\item{nCR}{\code{integer}. Length of vector with crossover probabilities for parameter subspace sampling (default: 3).}

\item{p_g}{\code{numeric}. Probability for gamma, the jump rate, being equal to 1. Default: 0.2.}

\item{beta0}{\code{numeric}. Reduce jump distance, e.g. if the average acceptance rate is low (less than 15 \%).
\code{0 < beta0 <= 1}. Default: 1 (i.e. jump distance is not adjusted).}

\item{thin}{\code{integer}. Thinning to be applied to output in case of large \code{t}. See below.}

\item{ncores}{\code{integer} specifying the number of cores to be used for parallel \code{pdf} evaluation
using the \code{\link[doMC]{doMC}} package (Linux only!). A value > 1 is only useful for complex \code{pdf}.
Default: 1.}

\item{verbose}{\code{logical}. Print progress bar to console? Default: TRUE.}
}
\value{
\code{list} with named elements:

\emph{chain}: a t/thin-by-d-by-nc array of parameter realisations for each iteration and Markov chain;

\emph{density}: a t/thin-by-nc matrix of densities computed by \code{pdf} at each iteration for each Markov chain;

\emph{runtime}: time of function execution in seconds;

\emph{outlier}: a list with t vectors of outlier indices in nc (value of 0 means no outliers);

\emph{AR}: a t/thin-by-nCR matrix giving the acceptance rate for each sample number and crossover value
(first element is NA due to computational reasons);

\emph{CR}: a t/thin-by-nCR matrix giving the selection probability for each sample number and crossover value
(first element is NA due to computational reasons);

\emph{R_stat}: a t/thin-by-d matrix giving the Gelman-Rubin convergence diagnostic (first two elements are NA
due to computational reasons).
}
\description{
Differential Evolution Adaptive Metropolis (DREAM) algorithm
}
\details{
To understand the notation (e.g. what is lambda, nCR etc.), have a look at Sect. 3.3
of the reference paper (see below).
}
\references{
Code from 'Algorithm 5' and 'Algorithm 6' of:

Vrugt, J. A.: "Markov chain Monte Carlo simulation using the DREAM software package:
Theory, concepts, and MATLAB implementation." Environmental Modelling & Software, 2016, 75, 273 -- 316,
\url{http://dx.doi.org/10.1016/j.envsoft.2015.08.013}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
